Move valid_to_reply_to? to RawEmail
`RawEmail` owns the `Mail::Message` instance that we interrogate to
calculate the value, so move the method there as it is the authoritative
source.

This commit required some of the attachment parsing specs to test
against a valid `RawEmail` instead of stubbed `Mail::Message` instance,
because `parse_raw_email!` is now checking against its `RawEmail`, which
currently must be persisted in order to establish whether it is
`valid_to_reply_to?`. There’s a load of duplication in this, but this
can be reduced by moving all these specs in to a single block describing
`#extract_attachments!`.

This also adds some supporting methods to `RawEmail` that are necessary
for `valid_to_reply_to?`, but these have been made private until we can
extract them from `IncomingMessage`.